consolidate to new core app; small migration updates
