+++ ohmg/georeference/views.py
@@ -9,7 +9,7 @@ from django.shortcuts import render, get_object_or_404
 from django.views import View
 from django.http import JsonResponse, HttpResponseBadRequest
 
-from ohmg.context_processors import generate_ohmg_context
+from ohmg.core.context_processors import generate_ohmg_context
 from ohmg.georeference.tasks import (
     run_preparation_session,
     run_georeference_session,
