+++ ohmg/settings.py
@@ -115,8 +115,8 @@ TEMPLATES = [
         "django.contrib.auth.context_processors.auth",
         "django.contrib.messages.context_processors.messages",
         "django.contrib.auth.context_processors.auth",
-        "ohmg.context_processors.navbar_footer_params",
-        "ohmg.context_processors.site_info",
+        "ohmg.core.context_processors.navbar_footer_params",
+        "ohmg.core.context_processors.site_info",
         "pinax_theme_bootstrap.context_processors.theme",
       ],
       "debug": DEBUG,
